style: 코드 경로 정리
