+++ question/generator.py
@@ -1,15 +1,18 @@
-import os
 import json
+import os
 from typing import List, Optional
 
 import openai
 from pydantic import BaseModel
 
-from config import AZURE_ENDPOINT, OPENAI_API_KEY, OPENAI_API_VERSION
+from question.config import AZURE_ENDPOINT, OPENAI_API_KEY, OPENAI_API_VERSION
 
 GPT_MODEL = "gpt-4o"
 GPT_INPUT_PRICE = 5 * 0.000001
 GPT_OUTPUT_PRICE = 15 * 0.000001
+
+QUESTION_ROOT = os.path.abspath(os.path.dirname(__file__))
+PROMPT_PATH = os.path.join(QUESTION_ROOT, "prompts", "question_generation_instruction.md")
 
 
 def to_user_message(text: str, image_urls: Optional[List[str]] = None):
@@ -73,6 +76,7 @@ class GptResponse(BaseModel):
 class QuestionGenerator:
     question_types = ['형태', '연산', '공통점, 차이점', '추천', '형식화 (리스트, 표)', '난이도, 언어', '연관 짓기 (1)', '연관 짓기 (2)']
     one_images_question_types = ['형태', '연산', '추천', '형식화 (리스트, 표)', '난이도, 언어']
+    prompt_path = PROMPT_PATH
 
     def __init__(self):
         self.openai_client = openai.AzureOpenAI(
@@ -81,15 +85,16 @@ class QuestionGenerator:
             api_version=OPENAI_API_VERSION,
         )
 
+    def get_instruction_prompt(self):
+        with open(self.prompt_path, "r") as f:
+            return f.read()
+
     def generate(self, image_urls: List[str]) -> GptResponse:
-        instruction_path = os.path.join("prompts", "question_generation_instruction.md")
-        with open(instruction_path, "r") as f:
-            instruction_prompt = f.read()
-
         user_prompt = "다음 이미지(들)에 대해서 사용자가 할 만한 질문들을 생성해주세요. 부탁드립니다."
-
-        instruction_message = to_system_message(instruction_prompt)
         user_command_message = to_user_message(user_prompt, image_urls)
+
+        instruction_prompt = self.get_instruction_prompt()
+        instruction_message = to_system_message(instruction_prompt)
 
         response = self.openai_client.beta.chat.completions.parse(
             model=GPT_MODEL,
